Updating author names.
I'm adding our names and affiliation.

The affiliation is screwed up in the output PDF because of a quirk
with the way that the AIP template handles commas in the affiliation
field, but that's hopefully something we don't have to mess with.
They can fix that when they typeset the thing for publication.